Use virtualenv for building
So the virtualenv-does-not-belong-in-container argument is null and
void! (SamYaple eyes ayoung...)

Due to the deps we have we end up installing python-requests from the
package manager. This leads to `pip wheel` failing to build because the
installed version of requests is too old. A venv solves this problem
(and removes the need for hash -r).

A venv should *not* be needed for project containers since the deps are
isolated. But should deps interfere we may be forced into venvs in the
project containers later. For now, it is isolated to builds.

Change-Id: Id38a115ba925fd6930789ed65db64f6e3709af3e
Depends-On: I453f8ecbb41cda3bb68ebc7ae5f5198ca9c9d772